read_6

+++ 1410002/update.c
@@ -4,7 +4,7 @@ void UpdateEvent(EVENT** root)
 {
 	int chk;
 	int i;
-	long int tmp;
+	long long int tmp;
 	EVENT* node = root;
 	printf("일정을 수정 할 수 있습니다. \n");
 	printf("어떤 일정을 수정하시겠습니까? \n");
